requirements.txt
Jinja2==3.0.3
langchain==0.0.354
numpy==1.24.4
python-dotenv==1.0.0
openai==1.7.2
chromadb==0.3.25